moved create event handleSubmit function to EventForm
